requirements.txt
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
exceptiongroup==1.1.3
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.2
oauth2client==4.1.3
opencv-python==4.8.0.76
outcome==1.2.0
pillar-youtube-upload==0.3.0
Pillow==10.0.0
proglog==0.1.10
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytube==15.0.0
requests==2.31.0
rsa==4.9
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0
youtube-channel-scraper==0.1.3
